Knowledge base Spring transition
This is the implementation of the knowledge base functionality that the old FixIT3150 system included.

The main change to the view of the users is an added 'Knowledge Base' page. This page is very similar to the 'Tickets' page, but filters tickets/articles shown so that only tickets with 'Knowledge Base' state are shown.

The changes added by this commit are:

- Added Knowledge Base JSP page.

- Added Knowledge Base link element to nav bar.

- Added method in TicketController for serving Knowledge Base page.

- Added method in TicketDao for fetching knowledge base articles.

- Changed the url pathname conditions in script.js from "/article" to "/knowledgeBase".

- Very minor formatting changes by my IDE.